Rewriting exoplanet beast AI to use an ai datum and nutrition.

+++ code/modules/mob/living/simple_animal/hostile/retaliate/exoplanet.dm
@@ -1,6 +1,6 @@
 
 /mob/living/simple_animal/hostile/retaliate/beast
-	var/hunger = 0
+	ai = /datum/ai/beast
 	var/list/prey = list()
 
 /mob/living/simple_animal/hostile/retaliate/beast/ListTargets(var/dist = 7)
@@ -12,34 +12,36 @@
 				var/mob/M = W.resolve()
 				if(M)
 					. |= M
-		else if(hunger > 500) //time to look for some food
+		else if(get_nutrition() < get_max_nutrition() * 0.75) //time to look for some food
 			for(var/mob/living/L in view(src, dist))
 				if(!attack_same && L.faction != faction)
 					prey |= weakref(L)
 
-/mob/living/simple_animal/hostile/retaliate/beast/Life()
-	. = ..()
-	if(!.)
-		return FALSE
-	hunger++
-	if(hunger < 100) //stop hunting when satiated
-		prey.Cut()
-	else if(stat == CONSCIOUS)
-		for(var/mob/living/simple_animal/S in range(src,1))
-			if(S == src)
-				continue
-			if(S.stat != DEAD)
-				continue
-			visible_message(SPAN_DANGER("\The [src] consumes the body of \the [S]!"))
-			var/turf/T = get_turf(S)
-			var/obj/item/remains/xeno/X = new(T)
-			X.desc += "These look like they belong to \a [S.name]."
-			hunger = max(0, hunger - 5*S.maxHealth)
-			if(prob(5))
-				S.gib()
-			else
-				qdel(S)
-			return
+/datum/ai/beast
+	expected_type = /mob/living/simple_animal/hostile/retaliate/beast
+
+/datum/ai/beast/do_process(time_elapsed)
+	var/mob/living/simple_animal/hostile/retaliate/beast/beast = body
+	var/nut = beast.get_nutrition()
+	var/max_nut = beast.get_max_nutrition()
+	if(nut > max_nut * 0.75 || beast.incapacitated())
+		beast.prey.Cut()
+		return
+	for(var/mob/living/simple_animal/S in range(beast,1))
+		if(S == beast)
+			continue
+		if(S.stat != DEAD)
+			continue
+		beast.visible_message(SPAN_DANGER("\The [beast] consumes the body of \the [S]!"))
+		var/turf/T = get_turf(S)
+		var/obj/item/remains/xeno/X = new(T)
+		X.desc += "These look like they belong to \a [S.name]."
+		beast.adjust_nutrition(5 * S.maxHealth)
+		if(prob(5))
+			S.gib()
+		else
+			qdel(S)
+		break
 
 /mob/living/simple_animal/proc/name_species()
 	set name = "Name Alien Species"
